chore(EMS-606): reusable generateSummaryListRows function for all summary lists
