+++ src/ui/server/helpers/summary-lists/answers-summary-list.ts
@@ -1,7 +1,7 @@
-import { FIELDS, LINKS, PAGES } from '../../content-strings';
+import { FIELDS, PAGES } from '../../content-strings';
 import { FIELD_IDS, ROUTES } from '../../constants';
-import getKeyText from './get-key-text';
-import { AnswersContent, AnswersFieldGroups, SummaryListItem, SummaryListItemData } from '../../../types';
+import generateSummaryListRows from './generate-summary-list-rows';
+import { AnswersContent, AnswersFieldGroups } from '../../../types';
 
 const {
   BUYER_COUNTRY,
@@ -157,42 +157,6 @@ const generateFieldGroups = (answers: AnswersContent) => {
 };
 
 /**
- * generateSummaryListRows
- * Map an array of fields with values for govukSummaryList component
- * @param {Array} Array of fields and answers
- * @returns {Array} Array of fields/answers in govukSummaryList data structure
- */
-const generateSummaryListRows = (fields: Array<SummaryListItemData>): Array<SummaryListItem> =>
-  fields.map((field: SummaryListItemData): SummaryListItem => {
-    const mapped = {
-      key: {
-        text: getKeyText(FIELDS, field.id),
-        classes: `${field.id}-key`,
-      },
-      value: {
-        text: field.value.text,
-        classes: `${field.id}-value`,
-      },
-      actions: {
-        items: [],
-      },
-    } as SummaryListItem;
-
-    if (field.renderChangeLink) {
-      mapped.actions.items.push({
-        href: field.href,
-        text: LINKS.CHANGE,
-        visuallyHiddenText: getKeyText(FIELDS, field.id),
-        attributes: {
-          'data-cy': `${field.id}-change-link`,
-        },
-      });
-    }
-
-    return mapped;
-  });
-
-/**
  * answersSummaryList
  * Create multiple groups with govukSummaryList data structure
  * @param {Object} All answers/submitted data in a simple object.text structure
@@ -215,4 +179,4 @@ const answersSummaryList = (answersContent: AnswersContent) => {
   return summaryList;
 };
 
-export { generateFieldGroups, generateSummaryListRows, answersSummaryList };
+export { generateFieldGroups, answersSummaryList };
