+++ src/ui/server/helpers/summary-lists/quote-summary-list.test.ts
@@ -1,9 +1,9 @@
-import { generateFields, generateSummaryListRows, quoteSummaryList } from './quote-summary-list';
+import { generateFields, quoteSummaryList } from './quote-summary-list';
 import mapQuoteToContent from '../data-content-mappings/map-quote-to-content';
-import { LINKS, QUOTE_TITLES } from '../../content-strings';
+import { QUOTE_TITLES } from '../../content-strings';
 import { FIELD_IDS, FIELD_VALUES, ROUTES } from '../../constants';
+import generateSummaryListRows from './generate-summary-list-rows';
 import { mockQuote } from '../../test-mocks';
-import { SummaryListItemData } from '../../../types';
 
 const { BUYER_COUNTRY, CONTRACT_VALUE, MAX_AMOUNT_OWED, MULTI_POLICY_LENGTH, PERCENTAGE_OF_COVER, POLICY_LENGTH, QUOTE, SINGLE_POLICY_LENGTH } = FIELD_IDS;
 
@@ -227,78 +227,6 @@ describe('server/helpers/summary-lists/quote-summary-list', () => {
         };
 
         expect(expectedField).toEqual(expected);
-      });
-    });
-  });
-
-  describe('generateSummaryListRows', () => {
-    const expectedObjBase = (field: SummaryListItemData) => ({
-      classes: 'ukef-white-text',
-      key: {
-        text: field.title,
-        classes: `${field.id}-key govuk-!-width-one-half`,
-      },
-      value: {
-        classes: `${field.id}-value`,
-      },
-      actions: {
-        items: [],
-      },
-    });
-
-    it('returns an array of objects mapped to quote content', () => {
-      const mockQuoteContent = mapQuoteToContent(mockQuote);
-
-      const fields = generateFields(mockQuoteContent);
-
-      const result = generateSummaryListRows(fields);
-
-      const expectedObj = (field: SummaryListItemData) => {
-        const initObj = expectedObjBase(field);
-
-        const mapped = {
-          ...initObj,
-          value: {
-            ...initObj.value,
-            text: field.value.text,
-          },
-        };
-        return mapped;
-      };
-
-      expect(result).toBeInstanceOf(Array);
-
-      const expected = expectedObj(fields[2]);
-      expect(result[2]).toEqual(expected);
-    });
-
-    describe('when a field has renderChangeLink', () => {
-      it('should add a link to action.itmes', () => {
-        const mockField = {
-          id: 'mock',
-          title: 'Test',
-          value: {
-            text: 'mock',
-          },
-          renderChangeLink: true,
-          href: '/page#field-label',
-        };
-
-        const result = generateSummaryListRows([mockField]);
-
-        const expected = [
-          {
-            href: mockField.href,
-            text: LINKS.CHANGE,
-            visuallyHiddenText: mockField.title,
-            attributes: {
-              'data-cy': `${mockField.id}-change-link`,
-            },
-            classes: 'ukef-white-text govuk-link--no-visited-state',
-          },
-        ];
-
-        expect(result[0].actions.items).toEqual(expected);
       });
     });
   });
